Fixing the find method. Not sure about the rest, tough.

+++ GrafoVis/polygons.pde
@@ -36,11 +36,11 @@ ArrayList get_polys() {
 }
 
 class Point {
-	public int x, y;
+	public float x, y;
 	public Polygon polygon;	
 	public Edge prev, next;
 	
-	public Point(int x, int y) {
+	public Point(float x, float y) {
 		this.x = x;
 		this.y = y;
 		this.prev = null;
